pyproject.toml
[project]
name = "big-query"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.43.0",
    "google-cloud-bigquery>=3.38.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.23.3",
]
